tools/code-jeeves/.out/emailClient/src/writeEmail7.test.ts


// Import writeEmail7 function from 'writeEmail7.ts'
import writeEmail7 from './writeEmail7';

// Testing
describe('writeEmail7', () => {
  // Test 1: writeEmail7 should be a function
  it('should be a function', () => {
    expect(typeof writeEmail7).toBe('function');
  });

  // Test 2: writeEmail7 should return a Promise
  it('should return a Promise', () => {
    expect(writeEmail7()).toBeInstanceOf(Promise);
  });

  // Test 3: writeEmail7 should write an email
  it('should write an email', async () => {
    // Arrange
    const email = await writeEmail7();

    // Act
    const emailBody = email.body;

    // Assert
    expect(emailBody).toBeTruthy();
  });
});
